test: isolate matrix binding harness to tests
