+++ src/channels/plugins/contracts/registry.ts
@@ -1,17 +1,9 @@
-import fs from "node:fs";
-import os from "node:os";
-import path from "node:path";
 import { expect, vi } from "vitest";
 import {
   __testing as discordThreadBindingTesting,
   createThreadBindingManager as createDiscordThreadBindingManager,
 } from "../../../../extensions/discord/runtime-api.js";
 import { createFeishuThreadBindingManager } from "../../../../extensions/feishu/api.js";
-import {
-  createMatrixThreadBindingManager,
-  resetMatrixThreadBindingsForTests,
-} from "../../../../extensions/matrix/src/matrix/thread-bindings.js";
-import { setMatrixRuntime } from "../../../../extensions/matrix/src/runtime.js";
 import { createTelegramThreadBindingManager } from "../../../../extensions/telegram/runtime-api.js";
 import type { OpenClawConfig } from "../../../config/config.js";
 import {
@@ -183,14 +175,6 @@ function expectClearedSessionBinding(params: {
 
 const telegramDescribeMessageToolMock = vi.fn();
 const discordDescribeMessageToolMock = vi.fn();
-const sendMessageMatrixMock = vi.hoisted(() =>
-  vi.fn(async () => ({ messageId: "$matrix-contract", roomId: "!room:example" })),
-);
-
-vi.mock("fake-indexeddb/auto", () => ({}));
-vi.mock("../../../../extensions/matrix/src/matrix/send.js", () => ({
-  sendMessageMatrix: sendMessageMatrixMock,
-}));
 
 bundledChannelRuntimeSetters.setTelegramRuntime({
   channel: {
@@ -220,43 +204,6 @@ bundledChannelRuntimeSetters.setLineRuntime({
       resolveLineAccount: ({ cfg, accountId }: { cfg: OpenClawConfig; accountId?: string }) =>
         resolveLineAccount({ cfg, accountId }),
     },
-  },
-} as never);
-
-const matrixContractStateDirs = new Set<string>();
-const MATRIX_CONTRACT_STATE_DIR_ENV = "OPENCLAW_MATRIX_CONTRACT_STATE_DIR";
-const matrixAuth = {
-  accountId: "default",
-  homeserver: "https://matrix.example.org",
-  userId: "@ops:example.org",
-  accessToken: "matrix-contract-token",
-  deviceId: "DEVICEID",
-};
-
-function createMatrixContractEnv(): NodeJS.ProcessEnv {
-  const stateDir = fs.mkdtempSync(path.join(os.tmpdir(), "openclaw-matrix-contracts-"));
-  matrixContractStateDirs.add(stateDir);
-  return {
-    ...process.env,
-    [MATRIX_CONTRACT_STATE_DIR_ENV]: stateDir,
-  };
-}
-
-function cleanupMatrixContractState(): void {
-  for (const stateDir of matrixContractStateDirs) {
-    fs.rmSync(stateDir, { recursive: true, force: true });
-  }
-  matrixContractStateDirs.clear();
-}
-
-export function resetMatrixSessionBindingContractStateForTests(): void {
-  resetMatrixThreadBindingsForTests();
-  cleanupMatrixContractState();
-}
-
-setMatrixRuntime({
-  state: {
-    resolveStateDir: (env) => env[MATRIX_CONTRACT_STATE_DIR_ENV] ?? os.tmpdir(),
   },
 } as never);
 
@@ -770,62 +717,15 @@ export const sessionBindingContractRegistry: SessionBindingContractEntry[] = [
       placements: ["current", "child"],
     },
     getCapabilities: async () => {
-      await createMatrixThreadBindingManager({
-        accountId: "default",
-        auth: matrixAuth,
-        client: {} as never,
-        env: createMatrixContractEnv(),
-        idleTimeoutMs: 24 * 60 * 60 * 1000,
-        maxAgeMs: 0,
-        enableSweeper: false,
-      });
-      return getSessionBindingService().getCapabilities({
-        channel: "matrix",
-        accountId: "default",
-      });
+      throw new Error("Matrix session binding test harness not installed");
     },
     bindAndResolve: async () => {
-      await createMatrixThreadBindingManager({
-        accountId: "default",
-        auth: matrixAuth,
-        client: {} as never,
-        env: createMatrixContractEnv(),
-        idleTimeoutMs: 24 * 60 * 60 * 1000,
-        maxAgeMs: 0,
-        enableSweeper: false,
-      });
-      const service = getSessionBindingService();
-      const binding = await service.bind({
-        targetSessionKey: "agent:matrix:subagent:child-1",
-        targetKind: "subagent",
-        conversation: {
-          channel: "matrix",
-          accountId: "default",
-          conversationId: "$thread-1",
-          parentConversationId: "!room:example",
-        },
-        placement: "current",
-        metadata: {
-          label: "codex-matrix",
-        },
-      });
-      expectResolvedSessionBinding({
-        channel: "matrix",
-        accountId: "default",
-        conversationId: "$thread-1",
-        targetSessionKey: "agent:matrix:subagent:child-1",
-      });
-      return binding;
+      throw new Error("Matrix session binding test harness not installed");
     },
-    unbindAndVerify: unbindAndExpectClearedSessionBinding,
-    cleanup: async () => {
-      resetMatrixSessionBindingContractStateForTests();
-      expectClearedSessionBinding({
-        channel: "matrix",
-        accountId: "default",
-        conversationId: "$thread-1",
-      });
+    unbindAndVerify: async () => {
+      throw new Error("Matrix session binding test harness not installed");
     },
+    cleanup: () => {},
   },
   {
     id: "telegram",
